feat(studio): add asset image preloader, co-locate filter input components
