+++ packages/sanity/src/core/studio/components/navbar/search/components/filters/filter/inputs/asset/Asset.tsx
@@ -2,17 +2,22 @@ import {ChevronDownIcon, ImageIcon, SearchIcon, UndoIcon} from '@sanity/icons'
 import type {Asset, AssetFromSource, AssetSource} from '@sanity/types'
 import {Box, Button, Flex, Menu, MenuButton, MenuItem, Portal, Stack} from '@sanity/ui'
 import React, {useCallback, useEffect, useId, useMemo, useState} from 'react'
-import {Source} from '../../../../../../../../config'
-import {FileSource, ImageSource} from '../../../../../../../../form/studio/assetSource'
-import {useClient} from '../../../../../../../../hooks'
-import {DEFAULT_STUDIO_CLIENT_OPTIONS} from '../../../../../../../../studioClient'
-import {useSource} from '../../../../../../../source'
-import {useSearchState} from '../../../../contexts/search/useSearchState'
-import {OperatorInputComponentProps} from '../../../../definitions/operators/operatorTypes'
-import {AssetSourceError} from './asset/AssetSourceError'
-import {AssetPreview} from './imagePreview/AssetPreview'
+import styled from 'styled-components'
+import {Source} from '../../../../../../../../../config'
+import {FileSource, ImageSource} from '../../../../../../../../../form/studio/assetSource'
+import {useClient} from '../../../../../../../../../hooks'
+import {DEFAULT_STUDIO_CLIENT_OPTIONS} from '../../../../../../../../../studioClient'
+import {useSource} from '../../../../../../../../source'
+import {useSearchState} from '../../../../../contexts/search/useSearchState'
+import type {OperatorInputComponentProps} from '../../../../../definitions/operators/operatorTypes'
+import {AssetSourceError} from './AssetSourceError'
+import {AssetPreview} from './preview/AssetPreview'
 
 type AssetType = keyof Pick<Source['form'], 'file' | 'image'>
+
+const ContainerBox = styled(Box)`
+  width: min(calc(100vw - 40px), 320px);
+`
 
 export function SearchFilterAssetInput(type?: AssetType) {
   return function FieldInputAssetWithType({onChange, value}: OperatorInputComponentProps<Asset>) {
@@ -90,7 +95,7 @@ export function SearchFilterAssetInput(type?: AssetType) {
     const buttonText = value ? `Change ${type}` : `Select ${type}`
 
     return (
-      <Box style={{width: 'min(calc(100vw - 40px), 320px)'}}>
+      <ContainerBox>
         <Stack space={3}>
           {/* Asset source component */}
           {selectedAssetSource && AssetSourceComponent && (
@@ -176,7 +181,7 @@ export function SearchFilterAssetInput(type?: AssetType) {
             )}
           </Flex>
         </Stack>
-      </Box>
+      </ContainerBox>
     )
   }
 }
