+++ packages/sanity/src/core/studio/components/navbar/search/components/filters/filter/inputs/asset/preview/ImagePreview.tsx
@@ -1,0 +1,45 @@
+import type {Asset} from '@sanity/types'
+import {Card, Flex, Spinner} from '@sanity/ui'
+import React, {useCallback, useState} from 'react'
+import styled from 'styled-components'
+
+interface ImagePreviewProps {
+  asset: Asset
+}
+
+const Container = styled(Card)`
+  position: relative;
+  padding-bottom: 100%;
+`
+
+const Image = styled.img`
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  display: block;
+  object-fit: contain;
+`
+
+const SpinnerFlex = styled(Flex)`
+  height: 100%;
+  position: absolute;
+  width: 100%;
+`
+
+export function ImagePreview({asset}: ImagePreviewProps) {
+  const [loaded, setLoaded] = useState(false)
+  const imageUrl = `${asset.url}?h=800&fit=max`
+  const handleLoad = useCallback(() => setLoaded(true), [])
+  return (
+    <Container __unstable_checkered border>
+      {!loaded && (
+        <SpinnerFlex align="center" justify="center">
+          <Spinner />
+        </SpinnerFlex>
+      )}
+      <Image onLoad={handleLoad} src={imageUrl} />
+    </Container>
+  )
+}
