+++ packages/sanity/src/core/studio/components/navbar/search/components/filters/filter/inputs/date/Date.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
-import {OperatorInputComponentProps} from '../../../../definitions/operators/operatorTypes'
-import {CommonDateInput} from './date/CommonDate'
+import {OperatorInputComponentProps} from '../../../../../definitions/operators/operatorTypes'
+import {CommonDateInput} from './CommonDate'
 
 export function SearchFilterDateInput(props: OperatorInputComponentProps<string>) {
   return <CommonDateInput {...props} />
