+++ packages/sanity/src/core/studio/components/navbar/search/components/filters/filter/inputs/number/NumberRange.tsx
@@ -1,8 +1,8 @@
 import {Box, Flex, TextInput} from '@sanity/ui'
 import React, {ChangeEvent, useCallback, useState} from 'react'
-import {useSearchState} from '../../../../contexts/search/useSearchState'
-import {OperatorNumberRangeValue} from '../../../../definitions/operators/common'
-import {OperatorInputComponentProps} from '../../../../definitions/operators/operatorTypes'
+import {useSearchState} from '../../../../../contexts/search/useSearchState'
+import {OperatorNumberRangeValue} from '../../../../../definitions/operators/common'
+import {OperatorInputComponentProps} from '../../../../../definitions/operators/operatorTypes'
 
 export function SearchFilterNumberRangeInput({
   onChange,
