+++ packages/sanity/src/core/studio/components/navbar/search/components/filters/filter/inputs/reference/ReferenceAutocomplete.tsx
@@ -2,13 +2,13 @@ import type {ReferenceValue} from '@sanity/types'
 import {Autocomplete, Box, Flex, Popover, Text} from '@sanity/ui'
 import React, {forwardRef, ReactElement, useCallback, useId, useMemo, useRef, useState} from 'react'
 import styled from 'styled-components'
-import {useSchema} from '../../../../../../../hooks'
-import type {SearchableType, WeightedHit} from '../../../../../../../search'
-import {getPublishedId} from '../../../../../../../util'
-import {useSearchState} from '../../../contexts/search/useSearchState'
-import {useSearch} from '../../../hooks/useSearch'
-import {documentTypesTruncated} from '../../../utils/documentTypesTruncated'
-import {SearchResultItem} from '../../searchResults/item/SearchResultItem'
+import {useSchema} from '../../../../../../../../../hooks'
+import type {SearchableType, WeightedHit} from '../../../../../../../../../search'
+import {getPublishedId} from '../../../../../../../../../util'
+import {useSearchState} from '../../../../../contexts/search/useSearchState'
+import {useSearch} from '../../../../../hooks/useSearch'
+import {documentTypesTruncated} from '../../../../../utils/documentTypesTruncated'
+import {SearchResultItem} from '../../../../searchResults/item/SearchResultItem'
 
 interface SearchHit {
   hit: WeightedHit
