+++ packages/sanity/src/core/studio/components/navbar/search/components/filters/filter/inputs/string/String.tsx
@@ -1,7 +1,7 @@
 import {TextInput} from '@sanity/ui'
 import React, {ChangeEvent, useCallback} from 'react'
-import {useSearchState} from '../../../../contexts/search/useSearchState'
-import {OperatorInputComponentProps} from '../../../../definitions/operators/operatorTypes'
+import {useSearchState} from '../../../../../contexts/search/useSearchState'
+import {OperatorInputComponentProps} from '../../../../../definitions/operators/operatorTypes'
 
 export function SearchFilterStringInput({
   onChange,
