+++ packages/sanity/src/core/studio/components/navbar/search/components/filters/filter/inputs/string/StringList.tsx
@@ -3,11 +3,11 @@ import {isTitledListValue, StringOptions, TitledListValue} from '@sanity/types'
 import {Box, Button, Code, Flex, Menu, MenuButton, MenuItem, Stack, Text} from '@sanity/ui'
 import {capitalize, uniq} from 'lodash'
 import React, {useCallback, useId, useMemo} from 'react'
-import {useSchema} from '../../../../../../../../hooks'
-import {isNonNullable} from '../../../../../../../../util'
-import {useSearchState} from '../../../../contexts/search/useSearchState'
-import {OperatorInputComponentProps} from '../../../../definitions/operators/operatorTypes'
-import {getSchemaField} from '../../../../utils/getSchemaField'
+import {useSchema} from '../../../../../../../../../hooks'
+import {isNonNullable} from '../../../../../../../../../util'
+import {useSearchState} from '../../../../../contexts/search/useSearchState'
+import {OperatorInputComponentProps} from '../../../../../definitions/operators/operatorTypes'
+import {getSchemaField} from '../../../../../utils/getSchemaField'
 
 interface TitledListValueGroup extends Omit<TitledListValue<number | string>, 'title'> {
   title: (number | string)[]
